+++ src/app/release.ts
@@ -12,6 +12,7 @@ export class Release {
   image: string;
   keywords: string[];
   credit: string;
+  quality: string;
 
   constructor(
     artist: string,
@@ -23,7 +24,8 @@ export class Release {
     url: string,
     image: string,
     keywords: string[],
-    credit: string
+    credit: string,
+    quality: string = ''
   ) {
     this.releaseItem = new ReleaseItem(url, artist, title);
     this.label = label;
@@ -34,6 +36,7 @@ export class Release {
     this.image = image;
     this.keywords = keywords;
     this.credit = credit;
+    this.quality = quality;
   }
 
   get artist(): ReleaseArtist {
@@ -79,7 +82,8 @@ export class Release {
       tracks: this.tracks.map((track) => track.toStorageObject()),
       image: this.image,
       keywords: this.keywords,
-      credit: this.credit
+      credit: this.credit,
+      quality: this.quality
     };
   }
 
@@ -110,7 +114,8 @@ export class Release {
       obj.url,
       obj.image,
       obj.keywords,
-      obj.credit ?? ''
+      obj.credit ?? '',
+      obj.quality ?? ''
     );
   }
 }